added BseGConfig object that exports the BseGlobals values through its
Wed Dec 29 22:41:27 1999  Tim Janik  <timj@gtk.org>

        * bseglobals.[hc]: added BseGConfig object that exports the BseGlobals
        values through its parameter interface, and provides notification for
        when BseGlobals are locked/unlocked.

        * bseglobals.c (bse_globals_unlock): clear up hunk/chunk memory pool
        when BseGlobals are unlocked.

        * bseheart.c: squeezy squeezy, the GSource was stealing a BseHeart ref
        count that it didn't own.

        * bsepcmdevice-oss.c (bse_pcm_device_oss_update_state): grr, the es1371
        driver in 2.3.34 can report audio_buf_info.bytes greater than
        audio_buf_info.fragsize * audio_buf_info.fragstotal, which could
        cause int wrap arounds in calculations we perform.

svn path=/trunk/; revision=116

+++ bse/bsechunk.c
@@ -104,7 +104,7 @@ bse_nuke_hunk_allocs (void)
   i = BSE_MAX_N_TRACKS - 1;
   if (g_trash_stack_height (&hunk_heap[i]) != hunk_count)
     {
-      g_warning ("hunk count discrepancy (%+d) - bleeding memory...", ((gint) i) - hunk_count);
+      g_warning ("hunk count discrepancy (%+d) - bleeding memory...", ((gint) i) - ((gint) hunk_count));
       hunk_heap[i] = NULL;
     }
   while (hunk_heap[i])
@@ -321,7 +321,6 @@ bse_chunk_unref (BseChunk *chunk)
 
       chunk_count--;
     }
-
 }
 
 void
@@ -454,7 +453,7 @@ bse_chunk_complete_state (BseChunk *chunk)
       g_return_if_fail (chunk->hunk_filled == TRUE);
 
       for (i = 0; i < BSE_TRACK_LENGTH * chunk->n_tracks; i++)
-	values[i % chunk->n_tracks] += chunk->hunk[i];
+	values[i % chunk->n_tracks] += chunk->hunk[i]; /* FIXME : optimize */
       
       for (i = 0; i < chunk->n_tracks; i++)
 	if (values[i] > 0)
@@ -485,7 +484,7 @@ bse_chunk_complete_hunk (BseChunk *chunk)
 	}
 
       for (i = 0; i < BSE_TRACK_LENGTH * chunk->n_tracks; i++)
-	chunk->hunk[i] = chunk->state[i % chunk->n_tracks];
+	chunk->hunk[i] = chunk->state[i % chunk->n_tracks]; /* FIXME : optimize */
 
       chunk->hunk_filled = TRUE;
     }
